requirements.txt
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
google-auth==2.3.3
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
gunicorn==20.1.0
